Delay death marks for 2 seconds after point capture
This commit is a series of changes required to delay death marks
for a period directly after a point is captured. All death marks are
cleared instantly upon point capture, meaning that for two seconds, no
one can be marked for death by the point for any reason, even if they
are a member of the opposing team who recently stood on the point.

The death mark timing system is now fully handled by plugin timers, not
by in-game condition timers, as was done previously. This should cut
down on double-marks and other strange mark behavior considerably.

This also standardizes the naming convention of arrays of client
attributes, i.e. client_on_point, client_last_class, etc.