Upgrade LMS minicrits on kill to full crits on kill
Playtesting has shown that a minicrit boost for the Last Man Standing
does not provide enough of an incentive for the LMS to rush in and
finish the match, nor does it prove to be a challenge for the team that
outnumbers them.

Upgrading this boost to full critical hits has proven to be a much
better balance and makes the 1vAll showdown more interesting to play
and to watch.