Experimental 5cp gameplay changes
This commit adds an untested 5cp playstyle. If a 5cp map is loaded, the
plugin will automatically choose a "mid" point to fight over, and two
"last" points that can be captured for an immediate round win.

Player spawn points are based around the chosen "mid" point, which functions
like a normal point in Arena: Respawn.