Fix doors opening early on KOTH maps
This commit introduces a significant number of entity-related utilities
required to detect a "spawn" door from a regular func_door.

In the 40+ Arena and KOTH maps tested, the spawn door detection logic
has worked well, forcing the appropriate func_doors open on round
start, and leaving spawns vulnerable as they would be in an Arena
match.

For a more in-depth discussion of the spawn door detection logic, refer
to this post on the official Arena: Respawn Steam group discussion
board: http://goo.gl/7PJxea