Increase health pickups
This commit adds a new method for handling the amount of health present
in Arena and KOTH maps. Generally speaking, KOTH maps have too much
health for Arena: Respawn, and (some) Arena maps have too little.

The temporary solution to this problem had been to downgrade all health
packs (large to medium, medium to small) on KOTH maps using Stripper.
Now health pack size is determined based on proximity to the point and
whether or not it is judged to be in a "contested" area (approximately
equidistant between both spawns).

This change has playtested well in both pub and PUG play by providing
new points of gameplay focus besides the control point itself.