Avoid balancing Arena teams over the start of a round
In public mode, do not attempt to balance teams or pull players from
the Spectate queue if a round has just ended. This should avoid
conflicts with the internal Arena team assignment logic, i.e. if the
plugin picked a player from the Spectate queue in the same frame the
Arena logic was scrambling teams (which it does by putting all players
in the Spectate queue first, then assigning them to Red or Blue).