Move team respawn code to a separate Team_RespawnWithEffects method
To prepare for potential other methods of respawning players (e.g.
touching a flag in CTF), move the code responsible for the team respawn
effect into its own team-specific method.

Team_RespawnWithEffects will both respawn the players on the associated
team and produce the "respawn" or "warning" sound accordingly, unless
told to be silent. By default, the method is not silent.