Add experimental CTF Arena mode
This addition to the A:R library still requires a significant amount
more testing and tweaking, but can be considered a playable beta for
now.

In CTF:R, the flag spawns in a predetermined point in the middle of the
map. The plugin will NOT be able to calculate a suitable midpoint
automatically; the midpoint must be placed manually as an info_target
with the targetname "flagpos," either inside the map itself, or added
by another plugin (such as Stripper: Source).

The spawned flag enables itself according to the time configured by
tf_arena_override_cap_enable_time, and can be picked up by either team.
When a player picks up the flag, any dead teammates will respawn.
However, a player can only hold the flag for up to 10 seconds. After
the 10 second timer has ended, the flag will be force dropped, and the
player and their team will have to defend the flag where it lies.

To prevent stalemates or hiding players, a successful 50-second defense
of the intel will reset the game state, respawn both teams, and award
the defending team with overheal and a single CTF point. Should the
game continue without a full wipe on either side, the first team to 3
successful flag defenses wins.